+++ dataforge-context/src/commonMain/kotlin/space/kscience/dataforge/properties/MetaProperty.kt
@@ -1,11 +1,7 @@
 package space.kscience.dataforge.properties
 
 
-import space.kscience.dataforge.meta.Meta
-import space.kscience.dataforge.meta.ObservableMutableMeta
-import space.kscience.dataforge.meta.transformations.MetaConverter
-import space.kscience.dataforge.meta.transformations.nullableMetaToObject
-import space.kscience.dataforge.meta.transformations.nullableObjectToMeta
+import space.kscience.dataforge.meta.*
 import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.startsWith
@@ -18,14 +14,14 @@ public class MetaProperty<T : Any>(
 ) : Property<T?> {
 
     override var value: T?
-        get() = converter.nullableMetaToObject(meta[name])
+        get() = converter.readNullable(meta[name])
         set(value) {
-            meta[name] = converter.nullableObjectToMeta(value) ?: Meta.EMPTY
+            meta[name] = converter.convertNullable(value) ?: Meta.EMPTY
         }
 
     override fun onChange(owner: Any?, callback: (T?) -> Unit) {
         meta.onChange(owner) { name ->
-            if (name.startsWith(this@MetaProperty.name)) callback(converter.nullableMetaToObject(this[name]))
+            if (name.startsWith(this@MetaProperty.name)) callback(converter.readNullable(this[name]))
         }
     }